+++ demopod/lib/dialogs/about.dart
@@ -1,6 +1,6 @@
 /// about dialog for the app
 ///
-// Time-stamp: <Friday 2024-05-17 14:31:16 +1000 Graham Williams>
+// Time-stamp: <Thursday 2024-06-27 13:38:24 +1000 Graham Williams>
 ///
 /// Copyright (C) 2024, Software Innovation Institute, ANU.
 ///
@@ -32,28 +32,30 @@ import 'package:flutter/material.dart';
 Future<void> aboutDialog(BuildContext context) async {
   final appInfo = await getAppNameVersion();
 
-  showAboutDialog(
-    context: context,
-    applicationLegalese: '© 2024 Software Innovation Institute ANU',
-    applicationIcon:
-        const ImageIcon(AssetImage('assets/images/demopod_logo.png')),
-    applicationName: appInfo.name,
-    applicationVersion: appInfo.version,
-    children: [
-      const SizedBox(
-        width: 300, // Limit the width.
-        child: SelectableText('\nA key-value pair manager.\n\n'
-            'Key Pod is an app for managing your secure and private'
-            ' key-value data in your Solid Pod. The key-value pairs'
-            ' can store any kind of data, indexed by the key.\n\n'
-            'The app is written in Flutter and the open source code'
-            ' is available from github at https://github.com/anusii/demopod.'
-            ' You can try it out online at https://demopod.solidcommunity.au.\n\n'
-            'The concept for the app and images were generated by'
-            ' large language models.\n\n'
-            'Authors: Anuska Vidanage, Graham Williams, Jess Moore, Zheyuan Xu,'
-            ' Kevin Wang, Ninad Bhat, Dawei Chen.'),
+  // Fix the use_build_context_synchronously lint error.
+
+  if (context.mounted) {
+    showAboutDialog(
+      context: context,
+      applicationName:
+          '${appInfo.name[0].toUpperCase()}${appInfo.name.substring(1)}',
+      applicationVersion: appInfo.version,
+      applicationLegalese: '© 2024 Software Innovation Institute ANU',
+      applicationIcon: Image.asset(
+        'assets/images/demopod_logo.png',
+        width: 100,
+        height: 100,
       ),
-    ],
-  );
+      children: [
+        const SizedBox(
+          width: 300, // Limit the width.
+          child: SelectableText('\nA demostrator of SolidPod functionality.\n\n'
+              'Demopod is a demonstrator app for the solidpod package.'
+              ' It provides a collection of buttons to exhibit the different'
+              ' calabilities provided by solidpod.\n\n'
+              'Authors: Anuska Vidanage, Graham Williams, Dawei Chen.'),
+        ),
+      ],
+    );
+  }
 }
